Move summary module into reports/details directory
