+++ gsa/src/web/pages/reports/details/summary.js
